requirements.txt
playwright
httpx
python-dotenv
supabase
torch
transformers
pillow
numpy
pandas